requirements.txt
argparse==1.4.0
colorama==0.4.6
dashscope==1.20.3
opencv-python==4.10.0.84
pyshine==0.0.9
PyYAML==6.0.2rc1
requests==2.32.3
sounddevice==0.4.7
matplotlib==3.9.1
keras==3.4.1
tensorflow==2.17.0
soundfile==0.12.1
torch==2.4.0
torchvision==0.19.0
gymnasium==0.29.1
numpy==1.26.4
pocoui==1.0.94
tqdm==4.66.5
seaborn==0.13.2
snippet-uiautomator==1.1.1
mobly~=1.12.3
Appium-Python-Client~=4.0.1
uiautomator~=1.0.2
qrcode~=7.4.2

pillow~=10.4.0
paho-mqtt~=2.1.0
ultralytics~=8.2.83